Split adapter types to their own definition file (#8205)
* Split adapter types to their own definition file
* Sort & separate
